Revert "valgrind: Enable valgrinding of Python plugin."
Further changes made in Python 3.7 have broken previous valgrinding.
It seems better not to try valgrind Python.

This reverts commit 4ceb2ef37b96f8f9cd395f382f07c6bc008ced98.

+++ tests/test-python-exception.sh
@@ -39,6 +39,12 @@ if test ! -d "$SRCDIR"; then
     exit 1
 fi
 
+# Python has proven very difficult to valgrind, therefore it is disabled.
+if [ "$NBDKIT_VALGRIND" = "1" ]; then
+    echo "$0: skipping Python test under valgrind."
+    exit 77
+fi
+
 output=test-python-exception.out
 
 rm -f $output
